requirements.txt
browser-use>=0.1.0
openai>=1.3.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
gspread==5.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0